Extract the serviceRef and types to argo-cd-node package
