+++ plugins/backend/backstage-plugin-argo-cd-backend/package.json
@@ -51,6 +51,7 @@
   "dependencies": {
     "@backstage/backend-plugin-api": "backstage:^",
     "@backstage/config": "backstage:^",
+    "@roadiehq/backstage-plugin-argo-cd-node": "^1.0.0",
     "@types/express": "^4.17.6",
     "cross-fetch": "^3.1.4",
     "express": "^4.17.1",
